+++ src/component/index.ts
@@ -12,28 +12,29 @@ export * from './backdrop/index';
 export * from './chips/index';
 export * from './circular-picker/index';
 export * from './code-viewer/index';
-export * from './color-picker/index'; // slow
+export * from './color-picker/index';
 export * from './color-selector/index';
-export * from './content-editable/index'; // slow
-export * from './data-grid/index'; // very slow
-export * from './date-picker/index'; // slow
+export * from './content-editable/index';
+export * from './data-grid/index';
+export * from './date-picker/index';
 export * from './date-selector/index';
-export * from './dialog/index'; // slow
+export * from './dialog/index';
 export * from './dragdrop/index';
-export * from './dropdown/index'; // very slow
+export * from './dropdown/index';
 export * from './iframe/index';
 export * from './loaders/index';
 export * from './markdown/index';
 export * from './menu/index';
 export * from './message-box/index';
-export * from './monaco-editor/index'; // slow
+export * from './monaco-editor/index';
 export * from './mouse-dragdrop/index';
 export * from './range/index';
 export * from './select/index';
 export * from './snackbar/index';
 export * from './splitter/index';
+export * from './tag/index';
 export * from './templates/index';
-export * from './tiles/index'; // very slow
-export * from './tooltip/index'; // slow
+export * from './tiles/index';
+export * from './tooltip/index';
 export * from './tree-list/index';
-export * from './viewport/index'; // slow
+export * from './viewport/index';