+++ src/component/tag/index.ts
@@ -1,0 +1,33 @@
+/*
+ *  @license
+ *  Copyright Hôpitaux Universitaires de Genève. All Rights Reserved.
+ *
+ *  Use of this source code is governed by an Apache-2.0 license that can be
+ *  found in the LICENSE file at https://github.com/DSI-HUG/dejajs-components/blob/master/LICENSE
+ */
+
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { FormsModule } from '@angular/forms';
+import { MdIconModule, MdInputModule} from '@angular/material';
+import { DejaChipsModule } from '../chips/index';
+import { DejaTagComponent } from './tag.component';
+
+@NgModule({
+    declarations: [DejaTagComponent],
+    exports: [DejaTagComponent],
+    imports: [
+        CommonModule,
+        FlexLayoutModule,
+        FormsModule,
+        MdIconModule,
+        MdInputModule,
+        FlexLayoutModule,
+        DejaChipsModule,
+    ],
+    providers: [],
+})
+export class DejaTagModule { }
+
+export * from './tag.component';